+++ src/main/java/de/vette/idea/neos/errorReporting/TriggerExceptionAction.java
@@ -1,0 +1,13 @@
+package de.vette.idea.neos.errorReporting;
+
+import com.intellij.openapi.actionSystem.AnAction;
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import org.jetbrains.annotations.NotNull;
+
+public class TriggerExceptionAction extends AnAction {
+
+    @Override
+    public void actionPerformed(@NotNull AnActionEvent e) {
+        throw new RuntimeException("I'm an artificial exception!");
+    }
+}